+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add processor respone code and processor response text to authorization adjustments subfield in transaction response.
+
 ## 3.46.0
 * Allow payee ID to be passed in options params for transaction create
 * Add `merchant_id` alias to ConnectedMerchantStatusTransitioned and ConnectedMerchantPayPalStatusChanged Auth webhooks